UI/requirements.txt
flask==2.3.3
pywebview==4.3.3
pillow>=9.0.0
python-dotenv==1.0.0
requests==2.31.0
